app/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
openai>=1.40.0
openai-whisper>=20231117
